signbridge-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
google-cloud-aiplatform==1.43.0
google-cloud-speech==2.25.0
google-cloud-texttospeech==2.16.3
google-cloud-storage==2.14.0
python-dotenv==1.0.1
